wrote portfolio
